requirements.txt
Babel==2.3.4
blinker==1.4
docutils==0.12
Jinja2==2.8
lxml==4.4.2
Markdown==3.2.1
MarkupSafe==0.23
Pygments==2.6.1
PyYAML==3.12
Werkzeug==0.11.11
Pillow==4.3.0
beautifulsoup4==4.8.2
feedgen==0.9.0
